+++ x/staking/exported/exported.go
@@ -1,4 +1,4 @@
-package expected
+package exported
 
 import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
